requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
appnope==0.1.3
    # via ipython
argh==0.30.3
    # via -r requirements.in
asttokens==2.4.1
    # via stack-data
authlib==1.2.1
    # via -r requirements.in
black==23.10.1
    # via -r requirements.in
blinker==1.7.0
    # via
    #   flask
    #   flask-principal
blis==0.7.11
    # via thinc
catalogue==2.0.10
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2023.7.22
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   flask
    #   flask-shell-ipython
    #   typer
cloudpathlib==0.16.0
    # via weasel
codecov==2.1.13
    # via -r requirements.in
confection==0.1.3
    # via
    #   thinc
    #   weasel
coverage[toml]==7.3.2
    # via
    #   codecov
    #   pytest-cov
cryptography==41.0.5
    # via authlib
cymem==2.0.8
    # via
    #   preshed
    #   spacy
    #   thinc
decorator==5.1.1
    # via ipython
dnspython==2.4.2
    # via email-validator
email-validator==2.1.0.post1
    # via flask-security-too
executing==2.0.1
    # via stack-data
flake8==6.1.0
    # via -r requirements.in
flake8-import-order==0.18.2
    # via -r requirements.in
flask==2.3.3
    # via
    #   -r requirements.in
    #   flask-login
    #   flask-principal
    #   flask-security-too
    #   flask-shell-ipython
    #   flask-sqlalchemy
    #   flask-wtf
flask-login==0.6.3
    # via flask-security-too
flask-principal==0.4.0
    # via flask-security-too
flask-security-too==5.3.2
    # via -r requirements.in
flask-shell-ipython==0.5.1
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
flask-wtf==1.2.1
    # via flask-security-too
freezegun==1.2.2
    # via -r requirements.in
fuzzywuzzy==0.18.0
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
idna==3.4
    # via
    #   email-validator
    #   requests
importlib-resources==6.1.0
    # via flask-security-too
iniconfig==2.0.0
    # via pytest
ipython==8.17.2
    # via flask-shell-ipython
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jedi==0.19.1
    # via ipython
jinja2==3.1.2
    # via
    #   flask
    #   spacy
json-log-formatter==0.5.2
    # via -r requirements.in
langcodes==3.3.0
    # via spacy
levenshtein==0.23.0
    # via python-levenshtein
markupsafe==2.1.3
    # via
    #   flask-security-too
    #   jinja2
    #   werkzeug
    #   wtforms
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via flake8
mohawk==1.1.0
    # via -r requirements.in
murmurhash==1.0.10
    # via
    #   preshed
    #   spacy
    #   thinc
mypy-extensions==1.0.0
    # via black
numpy==1.26.1
    # via
    #   -r requirements.in
    #   blis
    #   pandas
    #   spacy
    #   thinc
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.2
    # via
    #   black
    #   gunicorn
    #   pytest
    #   spacy
    #   thinc
    #   weasel
pandas==2.1.2
    # via -r requirements.in
parso==0.8.3
    # via jedi
passlib==1.7.4
    # via flask-security-too
pathspec==0.11.2
    # via black
pexpect==4.8.0
    # via ipython
platformdirs==3.11.0
    # via black
pluggy==1.3.0
    # via pytest
preshed==3.0.9
    # via
    #   spacy
    #   thinc
prompt-toolkit==3.0.39
    # via ipython
psycopg2==2.9.9
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.11.1
    # via
    #   flake8
    #   flake8-import-order
pycountry==22.3.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic==2.4.2
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
pydantic-core==2.10.1
    # via pydantic
pyflakes==3.1.0
    # via flake8
pygments==2.16.1
    # via ipython
pytest==7.4.3
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-mock
pytest-cov==4.1.0
    # via -r requirements.in
pytest-mock==3.12.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   freezegun
    #   pandas
python-dotenv==1.0.0
    # via -r requirements.in
python-levenshtein==0.23.0
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via -r requirements.in
rapidfuzz==3.5.1
    # via levenshtein
redis==5.0.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   codecov
    #   requests-oauthlib
    #   spacy
    #   weasel
requests-oauthlib==1.3.1
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   mohawk
    #   python-dateutil
smart-open==6.4.0
    # via
    #   spacy
    #   weasel
spacy==3.7.2
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.5
    # via spacy
sqlalchemy==2.0.22
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.1
    # via -r requirements.in
srsly==2.4.8
    # via
    #   confection
    #   spacy
    #   thinc
    #   weasel
stack-data==0.6.3
    # via ipython
thinc==8.2.1
    # via spacy
toolz==0.12.0
    # via -r requirements.in
tqdm==4.66.1
    # via spacy
traitlets==5.13.0
    # via
    #   ipython
    #   matplotlib-inline
typer==0.9.0
    # via
    #   spacy
    #   weasel
typing-extensions==4.8.0
    # via
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
tzdata==2023.3
    # via pandas
urllib3==2.0.7
    # via requests
wasabi==1.1.2
    # via
    #   spacy
    #   thinc
    #   weasel
watchdog==3.0.0
    # via -r requirements.in
wcwidth==0.2.9
    # via prompt-toolkit
weasel==0.3.3
    # via spacy
werkzeug==3.0.1
    # via
    #   -r requirements.in
    #   flask
    #   flask-login
wtforms==3.1.1
    # via
    #   flask-security-too
    #   flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
